requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
drf-yasg==1.21.7
mysqlclient==2.2.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
Pillow==10.1.0
